handling null values

+++ backend/src/main/java/com/pessom/todo/toDoItem/ToDoItemService.java
@@ -18,21 +18,28 @@ public class ToDoItemService {
     }
 
     // Updating a todo item
-    public ToDoItem updateToDoItem(Long toDoItemId, ToDoItem updatedToDoItem){
-        try {
-         ToDoItem item = toDoItemRepository.findById(toDoItemId).get();
-         updatedToDoItem.setId(item.getId());
-         updatedToDoItem.setStatus(item.getStatus());   
-        } catch (Exception e) {
-            System.out.println("ToDo Item not found");
+    public ToDoItem updateToDoItem(Long toDoItemId, ToDoItem updatedToDoItem) {
+        if (toDoItemId != null && updatedToDoItem != null) {
+            try {
+                ToDoItem item = toDoItemRepository.findById(toDoItemId).get();
+                updatedToDoItem.setId(item.getId());
+                updatedToDoItem.setStatus(item.getStatus());
+            } catch (Exception e) {
+                System.out.println("ToDo Item not found");
+            }
+
+            return toDoItemRepository.save(updatedToDoItem);
+        } else {
+            return new ToDoItem();
         }
-        
-        return toDoItemRepository.save(updatedToDoItem);
+
     }
 
     // Deleting a todo item
     public void deleteToDoItem(Long toDoItemId) {
-        toDoItemRepository.deleteById(toDoItemId);
+        if (toDoItemId != null) {
+            toDoItemRepository.deleteById(toDoItemId);
+        }
     }
 
     // Getting all todo items
@@ -42,7 +49,11 @@ public class ToDoItemService {
 
     // Getting one todo item
     public ToDoItem getToDoItemById(Long toDoItemId) {
-        return toDoItemRepository.findById(toDoItemId).get();
+        if (toDoItemId != null) {
+            return toDoItemRepository.findById(toDoItemId).get();
+        } else {
+            return new ToDoItem();
+        }
     }
 
     // Getting all finished todo items
@@ -56,8 +67,8 @@ public class ToDoItemService {
     }
 
     // Toggle between finished and unfinished
-    public ToDoItem toggleFinishedAndUnfinished(ToDoItem item, Status status){
+    public ToDoItem toggleFinishedAndUnfinished(ToDoItem item, Status status) {
         item.setStatus(status);
         return toDoItemRepository.save(item);
-    } 
+    }
 }
